Adds pages, applied navbar to those pages
This commit adds pages to the site that were missing. This now brings the page count to what is evident in the wireframes and design documentation. I have also added the code that displays the navigation bar and active class, although the active class isn't showing properly at the moment.