Typewriter effect on about me
This commit primarily adds a Scrolling typewriter effect which shows elements about me after a static "Im a" page. Once the animations finish. I am Aiming to keep "Im a Keen Coder" on The site. However, this is quite buggy in this commit.

Secondly, This lays down the framework for the buttons on the contact page which work and actually direct to the quoted social Media Site.

--------

Critical Restore Point D